% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_batch_spreadsheet.R
\name{create_batch_spreadsheet}
\alias{create_batch_spreadsheet}
\title{Creates a Batch CSV for us in the Data Aggregation Tool (version 1+)}
\usage{
create_batch_spreadsheet(path, Zone, shape, outdir)
}
\arguments{
\item{path}{a character string. The path of the project}

\item{Zone}{a character string. The Zone number (in quotation marks)}

\item{shape}{a character string. The shapefile name to use in the batch spreadsheet}

\item{outdir}{a character string. The path for the output}
}
\value{
A data frame. The data frame with the information for the batch spreadsheet
}
\description{
Will take a project folder, and then create a batch spreadsheet for use in data aggregation
}
\examples{
create_batch_spreadsheet(path = "R:/Projects/HP00000 - Projects/HP00104 - Seagreen Phase 2 and 3/",Zone = "85", shape = "Seagreen_Phase_2_and_3_16_km_buffer_WGS84.shp",outdir = "C:/Temp/")
}
