% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAT_rename.R
\name{DAT_rename}
\alias{DAT_rename}
\title{Custom function to rename the MATLAB created shapefiles}
\usage{
DAT_rename(DAT)
}
\arguments{
\item{DAT}{A centcount file output from the MATLAB Data Aggregation Tool. It
will be a shapefile (.shp)}
}
\value{
renames fields to reduce to 10 character limit
}
\description{
Changes some of the usual problem field names so they comply with the 10
character limit of filetype Shapefile (.shp).
}
\examples{
DAT <- "R:\\\\Projects\\\\HP00000 - Projects\\\\HP00119 - Cierco Atlantic Cluster (Llyr)\\\\700 - Analysis Reports\\\\files for EIA"
DAT_rename(DAT)
}
