% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatLonToUTMEPSGCode.R
\name{LatLonToUTMEPSGCode}
\alias{LatLonToUTMEPSGCode}
\title{UTM zone finder}
\usage{
LatLonToUTMEPSGCode(lat, lon)
}
\arguments{
\item{lat}{The WGS84 latitude as a numeric value}

\item{lon}{The WGS84 longitude as a numeric value}
}
\value{
A numeric ESPG UTM code
}
\description{
Find UTM EPSG code from Latitude and Longitude coordinates (EPSG 4326 WGS84)
(vectorised)
Source: https://geocompr.robinlovelace.net/reproj-geo-data.html
Source: https://gis.stackexchange.com/questions/13291/computing-utm-zone-from-lat-long-point
}
\examples{
UTM <- LatLonToUTMEPSGCode(45.33145,-18.33557)


}
