% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_libs.R
\name{Load_Libs}
\alias{Load_Libs}
\title{Custom function to load a list of libraries}
\usage{
Load_Libs(LIBS)
}
\arguments{
\item{LIBS}{a character vector. A list of libraries in the form c("x", "y")}
}
\value{
installs and loads libraries
}
\description{
Reads a list of libraries and installs them if they aren't already
}
\examples{
LIBS <- c("sp", "sf", "rgeos")
Load_Libs(LIBS)
}
