% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Obs_v_CentCount.R
\name{Obs_v_CentCount}
\alias{Obs_v_CentCount}
\title{Summarise observations versus Cent Count}
\usage{
Obs_v_CentCount(
  Observations,
  Obs_Shape,
  CentCount,
  SPP_lookup,
  Behaviour = "All",
  Boundary_Buffer
)
}
\arguments{
\item{Observations}{A data frame or tibble. The observation spreadsheet to be
opened and compared}

\item{Obs_Shape}{Observations opened as an 'sf' object}

\item{CentCount}{the CentCount as an sf object}

\item{SPP_lookup}{The species lookup spreadsheet opened as a data frame. This
is normally found at "R:/Overhead/HA17000 - R Code/5 - Bootstrap_Shiny/Hidef_bootstrap/data/Species_Lookup.xlsx"}

\item{Behaviour}{One of "Sitting", "Flying" or "All". The behaviour associated
with the CentCount to be summarized.}

\item{Boundary_Buffer}{The boundary as an "sf" object that encompasses all the
observations in the cent count}
}
\value{
A data frame. Summaries of the cent count and observations
}
\description{
This will take the observation spreadsheet, clip it to a boundary,
summarize the information by species and then compare. Useful for quick
Observation spreadsheet checks
}
\examples{
Observations <- openxlsx::read.xlsx("D:/TEMP/19_SG23_HiDef_Zone85_M12_S01_alldays_19_Observations.xlsx")
Obs_Shape <- sf::st_as_sf(Observations,coords=c("Latitude","Longitude"),crs=4326)
CentCount <- sf::st_read("D:/TEMP/Zone85_M12_S01_D01_19_Output.shp")
SPP_lookup <- openxlsx::read.xlsx("R:/Overhead/HA17000 - R Code/5 - Bootstrap_Shiny/Hidef_bootstrap/data/Species_Lookup.xlsx")
Boundary_Buffer <- sf::st_read("D:/TEMP/Development_area_4kmBuffer_WGS84.shp")
Output <- Obs_v_CentCount(Observations,Obs_Shape,CentCount,SPP_lookup,Behaviour="All",Boundary_Buffer)
}
