% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fortify.R
\name{read.fortify}
\alias{read.fortify}
\title{Custom function to read and fortify a polygon}
\usage{
read.fortify(filepath, epsg = 4326, return.read = F)
}
\arguments{
\item{filepath}{a character string. Filepath to the shapefile}

\item{epsg}{an integer. The EPSG of the output}

\item{return.read}{a boolean. Should the shapefile also be returned as read}
}
\value{
Either a list (if return.read=T) with the shapefile and the fortified data frame.
or just the fortified data frame.
}
\description{
Reads a polygon shapefile and fortifies it for ggplot
}
\examples{
X_UTM30 <- read.fortify("Temp/Shapefile.shp",epsg=32630,return.read=T)
X_WGS84 <- read.fortify("Temp/Shapefile.shp",epsg=4326,return.read=F)
}
