% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Split_CentCount.R
\name{Split_CentCount}
\alias{Split_CentCount}
\title{CentCount splitter}
\usage{
Split_CentCount(CentCount, Behaviour = c("Sitting", "Flying", "All"))
}
\arguments{
\item{CentCount}{the CentCount as an sf object}

\item{Behaviour}{One of "Sitting", "Flying" or "All". The behaviour associated
with the CentCount to be summarized.}
}
\value{
A list of size = 2. Unapportioned species counts and apportioned counts
}
\description{
Splits and summarises cent count for apportioned and unapportioned species
}
\examples{
All <- sf::st_read("D:/TEMP/Zone85_M01_S02_D01_21_Output.shp")
CentOut_All <- Split_CentCount(CentCount=All,Behaviour="All")
Flying <- sf::st_read("D:/TEMP/Zone85_M01_S02_D01_21_Flying.shp")
CentOut_Flying <- Split_CentCount(CentCount=Flying,Behaviour="Flying")
Sitting <- sf::st_read("D:/TEMP/Zone85_M01_S02_D01_21_Sitting.shp")
CentOut_Sitting <- Split_CentCount(CentCount=Sitting,Behaviour="Sitting")
}
