% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_species.R
\name{summarise_species}
\alias{summarise_species}
\title{Summarise species data for internal reporting}
\usage{
summarise_species(path, species)
}
\arguments{
\item{path}{a character string. The path of the observation spreadsheet}

\item{Species}{a character string. Name of the species}
}
\value{
A matrix. The displacement matrix for the species and season of interest
}
\description{
Will take an observation spreadsheet and summarise by a specific species
}
\examples{
test <- summarise_species(path = "R://Projects//HP00000 - Projects//HP00113 - SimplyBlue Erebus//700 - Analysis Reports//Observation spreadsheets//FOR_BEHAVIOUR_Zone86_M10_19_to_M09_21_ALL_OBS.xlsx",species = "Guillemot")
}
