% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Theme_HiDef.R
\name{Theme_HiDEF}
\alias{Theme_HiDEF}
\title{HiDef theme for plotting.}
\usage{
Theme_HiDEF(
  base_family = "Gill Sans MT",
  base_size = 11.5,
  plot_title_family = base_family,
  plot_title_size = 18,
  plot_title_face = "bold",
  plot_title_margin = 10,
  subtitle_family = base_family,
  subtitle_size = 12,
  subtitle_face = "plain",
  subtitle_margin = 15,
  strip_text_family = base_family,
  strip_text_size = 12,
  strip_text_face = "plain",
  caption_family = base_family,
  caption_size = 9,
  caption_face = "italic",
  caption_margin = 10,
  axis_text_size = base_size,
  axis_title_family = subtitle_family,
  axis_title_size = 9,
  axis_title_face = "plain",
  axis_title_just = "rt",
  plot_margin = margin(30, 30, 30, 30),
  grid_col = "#cccccc",
  grid = TRUE,
  axis_col = "#cccccc",
  axis = FALSE,
  ticks = FALSE
)
}
\value{
A ggplot theme
}
\description{
Based on theme_ipsum(). For help with this theme, check documentation on hrbrthemes::theme_ipsum()
}
\examples{
X <- data.frame(x=runif(100),y=runif(100))
ggplot(X,aes(x=x,y=y))+geom_point()+Theme_HiDEF()
}
