% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transect_pwr.R
\name{transect.pwr}
\alias{transect.pwr}
\title{Power analysis function for transect design}
\usage{
transect.pwr(
  CC,
  Species,
  nseq = NULL,
  t.space,
  kseq = seq(0, -1, by = -0.1),
  nSim = 10000,
  plot = TRUE,
  by.density = FALSE,
  alternative = "two.sided"
)
}
\arguments{
\item{CC}{An sf object. The opened cent count shapefile to process}

\item{Species}{A vector. The vector of species (using the two letter code)
that will be tested. E.G., c("S_CX","S_SP","S_RH","S_GN","S_GD","S_CA")}

\item{nseq}{A vector (optional). The number of transects to test. If NULL
then the code will test a halving, a doubling, and 1.5 times the spacing
of the original transects.}

\item{t.space}{Numeric. A value representing the transect spacing in the
cent count.}

\item{kseq}{Numeric sequence. Defaults to seq(0,1,by=0.1). The vector of
effect sizes to test (i.e., the % of original population size for which
statistical power is to be tested).}

\item{nSim}{Number of simulations to run power analysis on.}

\item{plot}{Boolean. Whether or not to plot the output in the console. A plot
will be returned in the output list.}

\item{by.density}{Boolean. If TRUE, will use density rather than count}

\item{alternative}{indicates the alternative hypothesis and must be one of
"two.sided" (default), "less", or "greater". You can specify just the initial
letter of the value, but the argument name must be given in full. See
‘Details’ of stats::ks.test() for the meanings of the possible values.}
}
\value{
A list. Contains plotobj (the plot as a ggplot), powervals (the data
frame with power values), and samplesize (the sample sizes as a data frame)
}
\description{
Taking a single set of transects from a HiDef digital aerial survey
will compute the statistical power for detecting population change.
}
\examples{
CentCount <- "D:/Power_Analysis/Data/Zone113_M02_S01_21_Output/Zone113_M02_S01_21_Output-CentCount.shp"
## Read the CentCount and convert to a data frame
CC <- sf::st_read(CentCount)

## Create the sequences of the original population (effect size)
kseq <- seq(0,-1,by=-0.1)
## Create the number of transects to test
nseq <- c(21,43,64,86)
labs <- c("5km","2.5km","1.68km","1.25km")
names(labs) <- nseq
Species <- c("S_CX","S_SP","S_RH","S_GN","S_GD","S_CA")
poweranalysis <- transect.pwr(CC=CC,Species = Species,nseq = nseq,kseq = kseq,t.space = 2.5,nSim = 1000, plot = TRUE, alternative = "less")
}
